requirements.txt
onnxruntime
transformers
torch
torchaudio
-e ./3rd-Party/IndicTransToolkit
python-multipart
streamlit
cython==3.1.2
fastapi==0.115.13
indicnlp==0.0.1
langchain==0.3.25
langchain_huggingface==0.3.0
langchain_openai==0.3.24
langchain_qdrant==0.2.0
numpy==2.3.0
pandas==2.3.0
processor==0.10.0
pydantic==2.11.7
python-dotenv==1.1.0
qdrant_client==1.14.3
sacrebleu==2.5.1
setuptools==78.1.1
soundfile==0.13.1
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
transformers==4.52.4
uvicorn==0.34.3
bitsandbytes==0.42.0
accelerate==0.30.1
